Put Product case class into a DBModel file
