+++ app/models/APIModel.scala
@@ -1,11 +1,7 @@
 package models
 
 import io.circe.{Decoder, Encoder}
-import io.circe.generic.JsonCodec
-import io.circe.syntax._
 import io.circe.generic.semiauto.{deriveDecoder, deriveEncoder}
-
-import java.time.OffsetDateTime
 
 case class ServiceStatus(
     service: String,
@@ -29,18 +25,4 @@ object ProductFromJson {
     deriveDecoder[ProductFromJson]
   implicit val productFromJsonEncoder: Encoder[ProductFromJson] =
     deriveEncoder[ProductFromJson]
-}
-case class Product(
-    id: String,
-    name: String,
-    quantity: Int,
-    price: Double,
-    createdAt: OffsetDateTime,
-    updatedAt: OffsetDateTime
-)
-object Product {
-  implicit val productDecoder: Decoder[Product] =
-    deriveDecoder[Product]
-  implicit val productEncoder: Encoder[Product] =
-    deriveEncoder[Product]
 }
